pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-live-translations"
version = "0.1.0"
description = "In-browser translation editing for Django superusers"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "django>=4.2",
    "polib>=1.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
    "pytest-sugar",
    "pytest-cov",
    "ruff",
    "pyrefly",
    "django-stubs",
    "pre-commit",
    "zensical",
]
e2e = [
    "pytest",
    "pytest-django",
    "pytest-playwright",
]
bench = [
    "pytest",
    "pytest-django",
    "pytest-benchmark",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D",       # docstrings - not needed everywhere
    "ANN",     # type annotations - pyrefly handles this
    "COM812",  # trailing comma - conflicts with formatter
    "ISC001",  # implicit string concat - conflicts with formatter
    "ERA",     # commented-out code
    "FIX",     # fixme comments
    "TD",      # todo formatting
    "DJ008",   # __str__ on models - not always needed
    "RUF012",  # mutable class default - Django model/admin convention
    "ARG",     # unused arguments - Django overrides require fixed signatures
    "SLF001",  # private member access - needed for Django internals patching
    "TRY003",  # raise vanilla args - too strict for a library
    "EM101",   # raw string in exception - too strict
    "EM102",   # f-string in exception - too strict
    "PLC0415", # import outside top level - Django lazy imports pattern
]

[tool.ruff.lint.per-file-ignores]
"tests/unit/**" = ["S101", "S105", "S106", "S108", "PLR2004", "PLR0913", "SIM117", "INP001"]
"tests/e2e/**" = ["S101", "S105", "S106", "S108", "S603", "S607", "PLR2004", "PLR0913", "PLR1714", "C420", "INP001"]
"tests/benchmarks/**" = ["S101", "S108", "PLR2004", "PLR0913", "FBT001", "FBT003", "B007", "PERF401", "INP001", "T201"]
"**/migrations/**" = ["ALL"]
"example/**" = ["S105", "INP001"]
"src/live_translations/strings.py" = ["C901", "PLR0915"]

[tool.pytest.ini_options]
addopts = "--import-mode=importlib"

[tool.ruff.lint.isort]
known-first-party = ["live_translations"]

[tool.pyrefly]
project-includes = ["src", "tests/unit"]
python-version = "3.12"

[tool.coverage.run]
source = ["src/live_translations"]

[tool.coverage.report]
exclude_lines = [
    "if t.TYPE_CHECKING:",
    "pragma: no cover",
    "except ImportError",
]
fail_under = 90

[tool.hatch.build.targets.wheel]
packages = ["src/live_translations"]
